+++ knowit/rules/__init__.py
@@ -1,5 +1,5 @@
 from knowit.rules.audio import AtmosRule, AudioChannelsRule, DtsHdRule
-from knowit.rules.general import LanguageRule
+from knowit.rules.general import CommentaryRule, LanguageRule
 from knowit.rules.subtitle import ClosedCaptionRule, HearingImpairedRule
 from knowit.rules.video import ResolutionRule
 
@@ -7,6 +7,7 @@ __all__ = [
     'AtmosRule',
     'AudioChannelsRule',
     'ClosedCaptionRule',
+    'CommentaryRule',
     'DtsHdRule',
     'HearingImpairedRule',
     'LanguageRule',